+++ R/formatData.R
@@ -7,7 +7,8 @@ formatPheno <- function(pheno, colname = NULL, classes = NULL) {
 
     # If no rownames in pheno_data
     if(all( as.character(seq_len(6)) %in% rownames(pheno)[seq_len(6)])){
-        pheno <- as.data.frame(pheno[, -1, drop=FALSE], row.names = as.character(pheno[[colnames(pheno)[1]]]))
+        pheno <- as.data.frame(pheno[, -1, drop=FALSE], row.names =
+                                   as.character(pheno[[colnames(pheno)[1]]]))
     }
 
     # If colname not in pheno_data