+++ inst/NEWS.md
@@ -1,5 +1,6 @@
 # spaceRAT 0.99.0
 
+* "logged" -> "exprs"
 * Preparing package for Bioconductor submission
 * Write vignettes
 * Add tests